requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
kafka-python==2.0.2
python-dotenv==1.0.0
tensorflow==2.13.0
requests==2.31.0
mlflow==2.8.0
boto3==1.29.0
psycopg2-binary==2.9.9